ingest/requirements.txt
requests
protobuf
grpcio
kafka-python
python-dotenv
